+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
-* Add a Class feature kind. We emit these when we encounter classes that aren't anything more specific (i.e. not an element, mixin, ...).
+* Add a Class feature kind for describing all kinds of classes. This is a superclass of the existing elements and mixins.
 
 ## [2.0.0-alpha.38] - 2017-04-13